@@ -22,7 +22,7 @@
     "test:unit": "npm run test:all -- --grep @integration --invert",
     "test:integration": "npm run test:all -- --grep @integration",
     "test:watch": "npm run test:unit -- --watch",
-    "lint": "eslint 'lib' 'bin' --ext .js --ignore-pattern lib/template.js --config .eslintrc",
+    "lint": "eslint 'lib' 'bin' --ext .js --ext .ts --ignore-pattern lib/template.js --config .eslintrc",
     "lint:fix": "npm run lint -- --fix",
     "semver": "semver"
   },
@@ -40,29 +40,36 @@
   },
   "homepage": "https://github.com/deluan/contentful-migrate#readme",
   "devDependencies": {
+    "@types/dateformat": "^5.0.0",
+    "@types/eslint": "^8.4.2",
+    "@types/expect.js": "^0.3.29",
+    "@types/listr": "^0.14.4",
+    "@types/mkdirp": "^1.0.2",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^17.0.33",
+    "@types/rimraf": "^3.0.2",
+    "@types/yargs": "^17.0.10",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "eslint": "^8.15.0",
     "expect.js": "^0.3.1",
     "growl": "^1.10.5",
     "husky": "^0.14.3",
-    "mocha": "^10.0.0"
+    "mocha": "^10.0.0",
+    "typescript": "^4.6.4"
   },
   "dependencies": {
     "chalk": "^3.0.0",
     "contentful-management": "^7.11.0",
-    "contentful-migration": "^4.5.0",
+    "contentful-migration": "^4.9.0",
     "dateformat": "^3.0.3",
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-mocha": "^6.3.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^4.2.0",
-    "eslint-plugin-standard": "^4.0.0",
     "listr": "^0.14.3",
     "lodash.reduce": "^4.6.0",
-    "migrate": "^1.6.2",
+    "migrate": "^1.8.0",
     "mkdirp": "^0.5.1",
     "p-map": "^4.0.0",
     "rimraf": "^3.0.2",
+    "tsup": "^5.12.8",
     "yargs": "^15.3.0"
   }
 }
