@@ -1,12 +1,12 @@
-const expect = require('expect.js')
-const {
+import expect from 'expect.js'
+import {
   removeNullValues,
   rejectEmptyObjects,
   createChangeFieldControl,
   restructureFields,
   restructureContentTypeJson,
   jsonToScript
-} = require('../jsonToScript')
+} from '../jsonToScript'
 
 const TIME_OUT = 5000 // in milliseconds
 
