@@ -1,17 +1,19 @@
-const path = require('path')
-const fs = require('fs')
-const log = require('migrate/lib/log')
-const logError = require('migrate/lib/log').error
-const mkdirp = require('mkdirp')
-const dateformat = require('dateformat')
+import path from 'path'
+import fs from 'fs'
+import log, { error as logError } from 'migrate/lib/log'
+import mkdirp from 'mkdirp'
+import dateformat from 'dateformat'
 
-const camelToDash = string => string.replace(/([A-Z])/g, char => `-${char.toLowerCase()}`)
+const camelToDash = (string: string) => string.replace(/([A-Z])/g, char => `-${char.toLowerCase()}`)
 
-const createFile = (contentTypeId, fileContent, migrationsDirectory) => {
+export interface CreatedFile { contentTypeId: string, fileName: string }
+export default function createFile(
+  contentTypeId: string, fileContent: string, migrationsDirectory: string
+): Promise<CreatedFile> {
   const directory = path.join(migrationsDirectory, contentTypeId)
   return new Promise((resolve, reject) => {
     // Ensure migrations directory exists
-    mkdirp(directory, (makeDirectoryError) => {
+    mkdirp(directory).catch((makeDirectoryError) => {
       if (makeDirectoryError) {
         reject(logError(`🚨  Failed to create ${directory}`, makeDirectoryError))
       }
@@ -22,7 +24,7 @@ const createFile = (contentTypeId, fileContent, migrationsDirectory) => {
       const filePath = path.join(directory, fileName)
 
       // Write the template file
-      return fs.writeFile(filePath, fileContent, (writeFileError) => {
+      return fs.promises.writeFile(filePath, fileContent).catch((writeFileError: any) => {
         if (writeFileError) {
           reject(logError(`🚨  Failed to create ${directory}/${fileName}`, writeFileError))
         }
@@ -33,4 +35,3 @@ const createFile = (contentTypeId, fileContent, migrationsDirectory) => {
   })
 }
 
-module.exports = createFile
