@@ -1,10 +1,8 @@
-const eslint = require('eslint')
-
-const removeNullValues = (value) => {
+export const removeNullValues = (value: any): any => {
   if (value instanceof Array) {
     return value.map(removeNullValues)
   } else if (value instanceof Object) {
-    const prunedObject = {}
+    const prunedObject: any = {}
     Object.keys(value).forEach((key) => {
       if (value[key] === null || value[key] === undefined) {
         return
@@ -16,44 +14,40 @@ const removeNullValues = (value) => {
   return value
 }
 
-const createProp = ([key, value]) => `
+export const createProp = ([key, value]: [string, any]) => `
   .${key}(${JSON.stringify(removeNullValues(value))})`
 
-const rejectEmptyObjects = ([, value]) => {
+export const rejectEmptyObjects = ([, value]: [string, any]) => {
   const emptyArray = value.constructor === Array && value.length === 0
   const emptyObject = typeof value === 'object' && Object.keys(value).length === 0
   return value && !emptyArray && !emptyObject
 }
 
-const createField = (itemId, field) => `
+export const createField = (itemId: string, field: any) => `
   ${itemId}.createField("${field.id}")${''.concat(...Object.entries(field.props).filter(rejectEmptyObjects).map(createProp))};
 `
-
-const createChangeFieldControl = (itemId, field) => {
+export const createChangeFieldControl = (itemId: string, field: any) => {
   const { fieldId, widgetId, settings } = field
   const baseString = `${itemId}.changeFieldControl("${fieldId}", "builtin", "${widgetId}"`
   return settings ? `${baseString}, ${JSON.stringify(settings)});` : `${baseString});`
 }
-
-const createContentType = (item, editorInterface) => `
+export const createContentType = (item: any, editorInterface: any) => `
   const ${item.id} = migration.createContentType('${item.id}')${''.concat(...Object.entries(item.props).map(createProp))};
-  ${''.concat(...item.fields.map(field => createField(item.id, field)))}
-  ${editorInterface.map(field => createChangeFieldControl(item.id, field)).join('\n')}
+  ${''.concat(...item.fields.map((field: any) => createField(item.id, field)))}
+  ${editorInterface.map((field: any) => createChangeFieldControl(item.id, field)).join('\n')}
 `
-
-const createScript = (item, editorInterface) => `module.exports.description = "Create content model for ${item.props.name}";
+export const createScript = (item: any, editorInterface: any) => `module.exports.description = "Create content model for ${item.props.name}";
 
   module.exports.up = (migration) => {${createContentType(item, editorInterface)}};
 
   module.exports.down = migration => migration.deleteContentType("${item.id}");
 `
 
-const restructureFields = (field) => {
+export const restructureFields = (field: any) => {
   const { id, ...props } = field
   return { id, props }
 }
-
-const restructureContentTypeJson = item => ({
+export const restructureContentTypeJson = (item: any) => ({
   id: item.sys.id,
   props: {
     name: item.name,
@@ -62,23 +56,14 @@ const restructureContentTypeJson = item => ({
   },
   fields: item.fields.map(restructureFields)
 })
-
-const jsonToScript = (contentTypeJson, editorInterface) => {
+export const jsonToScript = (contentTypeJson: any, editorInterface: any) => {
   const restructuredJson = restructureContentTypeJson(contentTypeJson)
   const unformattedScript = createScript(restructuredJson, editorInterface)
-  const engine = new eslint.CLIEngine({
-    fix: true,
-    baseConfig: { extends: ['eslint-config-standard'] },
-    useEslintrc: false
-  })
-  return engine.executeOnText(unformattedScript).results[0].output
-}
-
-module.exports = {
-  removeNullValues,
-  rejectEmptyObjects,
-  createChangeFieldControl,
-  restructureFields,
-  restructureContentTypeJson,
-  jsonToScript
+  // const engine = new eslint.CLIEngine({
+  //   fix: true,
+  //   baseConfig: { extends: ['eslint-config-standard'] },
+  //   useEslintrc: false
+  // })
+  // return engine.executeOnText(unformattedScript).results[0].output
+  return unformattedScript
 }
