@@ -1,11 +1,18 @@
-module.exports.description = "<Put your description here>";
+// @ts-check
+module.exports.description = '<Put your description here>'
 
+/**
+ * @param {import("contentful-migration").default} migration
+ */
 module.exports.up = (migration) => {
   // Add your UP migration script here. See examples here:
   // https://github.com/contentful/migration-cli/tree/master/examples
-};
+}
 
+/**
+ * @param {import("contentful-migration").default} migration
+ */
 module.exports.down = (migration) => {
   // Add your DOWN migration script here. See examples here:
   // https://github.com/contentful/migration-cli/tree/master/examples
-};
+}
